+++ src/python/dxpy/bindings/dxworkflow.py
@@ -35,6 +35,8 @@ from ..compat import basestring
 ##############
 # DXWorkflow #
 ##############
+_workflow_required_keys = ['name', 'outputFolder']
+_workflow_stage_keys = ['id', 'name', 'executable', 'folder', 'input', 'executionPolicy', 'systemRequirements']
 
 def new_dxworkflow(title=None, summary=None, description=None, output_folder=None, init_from=None, **kwargs):
     '''
@@ -457,6 +459,12 @@ class DXWorkflow(DXDataObject, DXExecutable):
             ]
 
         return run_input
+
+    def _get_required_keys(self):
+        return _workflow_required_keys
+
+    def _get_stage_keys(self):
+        return _workflow_stage_keys
 
     def _run_impl(self, run_input, **kwargs):
         return DXAnalysis(dxpy.api.workflow_run(self._dxid, run_input, **kwargs)["id"])